feat(oidc-client): move fetch to oidc-client
